extract out `PoolInfo`
